Migrate AnnotationContext datatypes
